Fix g++-3.4 warnings.
	* mln/world/binary_2d/enlarge.hh (do_threshold): Fix.
	Upgrade doc.
	* mln/core/image/edge_image.hh: Likewise.
	* mln/core/dpoint.hh: Likewise.
	(dpoint): Fix implicit cast.
	* mln/metal/converts_to.hh: Upgrade doc.
	(converts_to): Add overloads to quiet g++-3.4.
	* mln/level: Remove directory.
	* mln/algebra/mat.hh: Upgrade doc.
	(inverse): Fix; use float instead of T.
	* mln/io/off/load.hh: Upgrade doc.
	(read_face_data): Fix missing scaling to 255.
	Add explicit conversion.
	* tests/core/image/graph_image.cc,
	* tests/core/image/edge_image.cc,
	* tests/core/image/line_graph_image.cc,
	* tests/core/image/vertex_image.cc: Upgrade doc.
	(X): New.
	(expected_fwd_nb, expected_bkd_nb): Update.
	Fix negative value stored as unsigned.


git-svn-id: https://svn.lrde.epita.fr/svn/oln/trunk@3946 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ milena/mln/world/binary_2d/enlarge.hh
@@ -30,7 +30,7 @@
 
 /// \file mln/world/binary_2d/enlarge.hh
 ///
-/// Enlarge 2^n times a binary image.
+/// \brief Enlarge 2^n times a binary image.
 ///
 /// \todo Make it much more generic.
 
@@ -97,7 +97,7 @@ namespace mln
 	int
 	do_threshold(float value)
 	{
-	  return 255.f * value;
+	  return static_cast<int>(255.f * value);
 	}
 
       } // end of namespace mln::world::binary_2d
